+++ examples/KafkaPublishExample.php
@@ -12,7 +12,6 @@ $broker = 'localhost:9092';
 //$conf->set('auto.offset.reset', 'largest');
 //
 //$consumer = new \RdKafka\KafkaConsumer($conf);
-
 // create producer
 $conf = new \RdKafka\Conf();
 $conf->set('bootstrap.servers', $broker);
@@ -21,9 +20,11 @@ $conf->set('queue.buffering.max.ms', 20);
 $producer = new \RdKafka\Producer($conf);
 $producer->addBrokers($broker);
 
-$adapter = new \Superbalist\PubSub\Kafka\KafkaPubSubAdapter($producer);
-
-for ($x = 0; $x < 10; $x++) {
-    echo "Publishing to tempTopic ".$x."\n";
-    $adapter->publish('tempTopi1c', $x);
+$adapter = new \milind\PubSub\Kafka\KafkaPubSubAdapter($producer);
+$topic = 'partition-test';
+while (true) {
+    for ($x = 0; $x < 100; $x++) {
+        echo "Publishing to tempTopic " . $x . "\n";
+        $adapter->publish($topic, $x, $x);
+    }
 }